Role permission set on button

+++ app/Http/Controllers/Auth/RoleController.php
@@ -5,10 +5,11 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use App\Services\Roles\RoleServices;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
+use Illuminate\Contracts\View\View;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\Gate;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -25,9 +26,11 @@ class RoleController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(): View
     {
-        return view('pages.roles.index');
+        Gate::authorize('app.roles.index');
+
+        return view('pages.roles.index', $this->services->accessPermissions());
     }
 
     /**
@@ -39,6 +42,8 @@ class RoleController extends Controller
      */
     public function getRoles(): LengthAwarePaginator
     {
+        Gate::authorize('app.roles.index');
+
         return $this->services->getRoles();
     }
 
@@ -49,6 +54,8 @@ class RoleController extends Controller
      */
     public function getPermissions(): Collection|array
     {
+        Gate::authorize('app.roles.create');
+
         return $this->services->permissions();
     }
 
@@ -57,6 +64,8 @@ class RoleController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        Gate::authorize('app.roles.create');
+
         return $this->services->validateStore()->store($request);
     }
 
@@ -65,6 +74,8 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id): JsonResponse
     {
+        Gate::authorize('app.roles.edit');
+
         return $this->services->validateUpdate($request, $id)->update($request, $id);
     }
 
@@ -73,6 +84,8 @@ class RoleController extends Controller
      */
     public function destroy(string $id): JsonResponse
     {
+        Gate::authorize('app.roles.delete');
+
         return $this->services->delete($id);
     }
 }
